requirements.txt
pdf2image
pytesseract
faiss-cpu
sentence-transformers
transformers
torch==2.5.1+cu121
python-docx
Pillow==11.0.0
numpy==2.1.2
